+++ packages/input/src/parts/index.ts
@@ -1,6 +1,7 @@
-export * from './StyledInput'
-export * from './PrefixSuffix'
 export * from './Addon'
+export * from './StyledInput'
+
+export * from './PrefixSuffix'
 export * from './LabelContainer'
 
 export * from './Attachments'