+++ packages/theme/__tests__/theme/theme/borders.ts
@@ -1,0 +1,5 @@
+import { Borders } from '@atlantis-lab/theme'
+
+export const borders: Borders = {
+  primary: '1px solid red',
+}