+++ packages/theme/__tests__/theme/theme/fontWeights.ts
@@ -1,0 +1,6 @@
+import { FontWeights } from '@atlantis-lab/theme'
+
+export const fontWeights: FontWeights = {
+  small: 1,
+  big: '1rem',
+}