+++ packages/theme/__tests__/theme/theme/fonts.ts
@@ -1,0 +1,5 @@
+import { Fonts } from '@atlantis-lab/theme'
+
+export const fonts: Fonts = {
+  primary: 'anyfont',
+}