+++ packages/theme/__tests__/theme/theme/radii.ts
@@ -1,0 +1,6 @@
+import { Radii } from '@atlantis-lab/theme'
+
+export const radii: Radii = {
+  small: 1,
+  big: '1rem',
+}