+++ packages/theme/__tests__/theme/theme/shadows.ts
@@ -1,0 +1,5 @@
+import { Shadows } from '@atlantis-lab/theme'
+
+export const shadows: Shadows = {
+  primary: 'anyshadow',
+}