+++ packages/theme/src/stories/theme/fonts.ts
@@ -1,0 +1,3 @@
+import { Fonts } from '../../theme'
+
+export const fonts: Fonts = {}