+++ packages/theme/CHANGELOG.md
@@ -3,6 +3,206 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.28 (2021-03-25)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.27 (2021-03-24)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.26 (2021-03-24)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.25 (2021-03-24)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.24 (2021-03-17)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.23 (2021-03-17)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.22 (2021-03-10)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.21 (2021-03-10)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.20 (2021-03-10)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.19 (2021-03-10)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.18 (2021-03-10)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.17 (2021-03-10)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.16 (2021-03-09)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.15 (2021-03-09)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.14 (2021-03-08)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.13 (2021-03-08)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.12 (2021-03-08)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.11 (2021-03-08)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.10 (2021-03-08)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.9 (2021-02-28)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.8 (2021-02-26)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.7 (2021-02-26)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.6 (2021-02-26)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.5 (2021-02-26)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
+## 0.0.4 (2021-02-26)
+
+**Note:** Version bump only for package @atlantis-lab/theme
+
+
+
+
+
 ## 0.0.3 (2021-02-26)
 
 **Note:** Version bump only for package @atlantis-lab/theme